feat: add ActivityLauncher with delegation pattern
